+++ MoneyFi-backend/MoneyFi-income/src/main/java/com/moneyfi/income/service/IncomeService.java
@@ -8,28 +8,28 @@ public interface IncomeService {
 
     public IncomeModel save(IncomeModel income);
 
-    List<IncomeModel> getAllIncomes(int userId);
+    List<IncomeModel> getAllIncomes(Long userId);
 
-    public List<IncomeModel> getAllIncomesByDate(int userId, int month, int year, boolean deleteStatus);
+    public List<IncomeModel> getAllIncomesByDate(Long userId, int month, int year, boolean deleteStatus);
 
-    public byte[] generateMonthlyExcelReport(int userId, int month, int year);
+    public byte[] generateMonthlyExcelReport(Long userId, int month, int year);
 
-    public List<IncomeModel> getAllIncomesByYear(int userId, int year, boolean deleteStatus);
+    public List<IncomeModel> getAllIncomesByYear(Long userId, int year, boolean deleteStatus);
 
-    public byte[] generateYearlyExcelReport(int userId, int year);
+    public byte[] generateYearlyExcelReport(Long userId, int year);
 
-    public List<Double> getMonthlyIncomes(int userId, int year);
+    public List<Double> getMonthlyIncomes(Long userId, int year);
 
-    public Double getTotalIncomeInMonthAndYear(int userId, int month, int year);
+    public Double getTotalIncomeInMonthAndYear(Long userId, int month, int year);
 
-    public Double getRemainingIncomeUpToPreviousMonthByMonthAndYear(int userId, int month, int year);
+    public Double getRemainingIncomeUpToPreviousMonthByMonthAndYear(Long userId, int month, int year);
 
     public boolean incomeUpdateCheckFunction(IncomeModel incomeModel);
 
     public boolean incomeDeleteCheckFunction(IncomeModel incomeModel);
 
-    public IncomeModel updateBySource(int id, IncomeModel income);
+    public IncomeModel updateBySource(Long id, IncomeModel income);
 
-    public boolean deleteIncomeById(int id);
+    public boolean deleteIncomeById(Long id);
 
 }
