+++ MoneyFi-backend/MoneyFi-user-details/src/main/java/com/moneyfi/user/service/profile/ProfileService.java
@@ -4,9 +4,9 @@ import com.moneyfi.user.model.ProfileModel;
 
 public interface ProfileService {
 
-    public ProfileModel save(int userId, ProfileModel profile);
+    public ProfileModel save(Long userId, ProfileModel profile);
 
-    public ProfileModel findByUserId(int userId);
+    public ProfileModel findByUserId(Long userId);
 
-    public String getNameByUserId(int userId);
+    public String getNameByUserId(Long userId);
 }
